+++ packages/origin-backend-client-mocks/src/OffChainDataSourceMock.ts
@@ -5,6 +5,7 @@ import {
     IUserClient,
     IDeviceClient,
     IOrganizationClient,
+    IInvitationClient,
     IRequestClient,
     IFilesClient,
     ICertificateClient,
@@ -16,6 +17,7 @@ import { ConfigurationClientMock } from './ConfigurationClientMock';
 import { UserClientMock } from './UserClientMock';
 import { DeviceClientMock } from './DeviceClientMock';
 import { OrganizationClientMock } from './OrganizationClientMock';
+import { InvitationClientMock } from './InvitationClientMock';
 import { CertificateClientMock } from './CertificateClientMock';
 import { CertificationRequestClientMock } from './CertificationRequestClientMock';
 
@@ -29,6 +31,8 @@ export class OffChainDataSourceMock implements IOffChainDataSource {
     deviceClient: IDeviceClient = new DeviceClientMock();
 
     organizationClient: IOrganizationClient = new OrganizationClientMock();
+
+    invitationClient: IInvitationClient = new InvitationClientMock();
 
     requestClient: IRequestClient = new RequestClient();
