+++ packages/origin-backend-client-mocks/src/OrganizationClientMock.ts
@@ -1,34 +1,19 @@
 import {
     IOrganizationInvitation,
     IUser,
-    OrganizationInvitationStatus,
     ISuccessResponse,
     OrganizationPostData,
     OrganizationStatus,
     OrganizationUpdateData,
-    OrganizationRole,
     Role,
     IOrganizationClient,
     IFullOrganization
 } from '@energyweb/origin-backend-core';
 
-interface ITmpUser {
-    id: number;
-    email: string;
-}
-
 export class OrganizationClientMock implements IOrganizationClient {
     private storage = new Map<number, IFullOrganization>();
 
-    private invitationStorage = new Map<number, IOrganizationInvitation>();
-
-    private userStorage: ITmpUser[] = [];
-
     private idCounter = 0;
-
-    private invitationCounter = 0;
-
-    private userCounter = 0;
 
     async getById(id: number): Promise<IFullOrganization> {
         return this.storage.get(id);
@@ -62,41 +47,6 @@ export class OrganizationClientMock implements IOrganizationClient {
         return Promise.resolve(organization);
     }
 
-    invite(email: string, role: OrganizationRole): Promise<ISuccessResponse> {
-        throw new Error('Method not implemented.');
-    }
-
-    inviteMocked(
-        email: string,
-        organization: IFullOrganization,
-        role: OrganizationRole,
-        sender: string
-    ): ISuccessResponse {
-        this.invitationCounter++;
-
-        const organizationInvitation: IOrganizationInvitation = {
-            id: this.invitationCounter,
-            email,
-            role,
-            organization,
-            status: OrganizationInvitationStatus.Pending,
-            sender,
-            createdAt: new Date()
-        };
-
-        this.invitationStorage.set(organizationInvitation.id, organizationInvitation);
-
-        this.userCounter++;
-        this.userStorage.push({
-            id: this.userCounter,
-            email
-        });
-
-        return {
-            success: true,
-            message: organizationInvitation.id.toString()
-        };
-    }
 
     getMembers(id: number): Promise<IUser[]> {
         throw new Error('Method not implemented.');
@@ -114,27 +64,8 @@ export class OrganizationClientMock implements IOrganizationClient {
         throw new Error('Method not implemented.');
     }
 
-    getInvitations(): Promise<IOrganizationInvitation[]> {
+    getInvitationsForOrganization(organizationId: number): Promise<IOrganizationInvitation[]> {
         throw new Error('Method not implemented.');
     }
 
-    getInvitationsToOrganization(organizationId: number): Promise<IOrganizationInvitation[]> {
-        throw new Error('Method not implemented.');
-    }
-
-    getInvitationsForEmail(email: string): Promise<IOrganizationInvitation[]> {
-        throw new Error('Method not implemented.');
-    }
-
-    acceptInvitation(invitationId: number): Promise<any> {
-        throw new Error('Method not implemented.');
-    }
-
-    rejectInvitation(id: number): Promise<any> {
-        throw new Error('Method not implemented.');
-    }
-
-    viewInvitation(id: number): Promise<any> {
-        throw new Error('Method not implemented.');
-    }
 }
